mostrando img en ruta /images/id_img

+++ src/controllers/image.js
@@ -1,68 +1,70 @@
-const path  = require('path');
-const { randomNumber} = require('../helpers/libs');
-const fs = require('fs-extra');
+const path = require("path");
+const { randomNumber } = require("../helpers/libs");
+const fs = require("fs-extra");
 
-const {Image} = require('../models/index');
+const { Image } = require("../models/index");
 
 const ctrl = {};
 
-ctrl.index  = (req, res) => {
+ctrl.index = async (req, res) => {
+
+const image = await Image.findOne({filename: {$regex :req.params.image_id}}).lean();
 
 
-
+console.log(image);
+  res.render('image',{image});
 };
 
-ctrl.create   =  async (req, res) => {
-
+ctrl.create = (req, res) => {
+  const saveImage = async () => {
     const imgUrl = randomNumber();
-    console.log(imgUrl);
-
-    const imageTempPath = req.file.path;
-
-
-//extension de la imagen
-    const ext = path.extname(req.file.originalname).toLowerCase();
-    const targetPath = path.resolve(`src/public/upload/${imgUrl}${ext}`)
-
-
-    if (ext === '.png' || ext === '.jpg' || ext === '.jpeg' || ext === '.gif') {
+    const images = await Image.find({filename: imgUrl})
+  
+    if (images.length > 0) {
+        saveImage();
         
-        await fs.rename(imageTempPath, targetPath);
-        const newImage = new Image({
+    }else{
+        // console.log(imgUrl);
+
+        const imageTempPath = req.file.path;
+        //extension de la imagen
+        const ext = path.extname(req.file.originalname).toLowerCase();
+        const targetPath = path.resolve(`src/public/upload/${imgUrl}${ext}`);
+    
+        if (ext === ".png" || ext === ".jpg" || ext === ".jpeg" || ext === ".gif") {
+          await fs.rename(imageTempPath, targetPath);
+          const newImg = new Image({
             title: req.body.title,
             filename: imgUrl + ext,
-            descripcion: req.body.descripcion
+            descripcion: req.body.descripcion,
+          });
+    
+          const ImageSaved = await newImg.save();
+           res.redirect(`./images/${imgUrl}`);
+        // res.send('work!');
 
-        });
+          // console.log(newImg);
+        } else {
+          await fs.unlink(imageTempPath);
+          res.status(500).json({ error: "Only Images are allowed" });
+        }
+    
+      };
 
-        const ImageSaved = await newImage.save();
-        console.log(newImage);
-
-    }else{
-        await fs.unlink(imageTempPath);
-        res.status(500).json({error: 'Only Images are allowed'});
     }
-
-
-res.send('funciona')
+    
+  saveImage();
+ 
 };
 
-
-
-ctrl.like   =  (req, res) => {
-
-    res.send('Index page');
-
+ctrl.like = (req, res) => {
+  res.send("Index page");
 };
-ctrl.comment   =  (req, res) => {
-
-    res.send('Index page');
-
+ctrl.comment = (req, res) => {
+  res.send("Index page");
 };
-ctrl.remove   =  (req, res) => {
-
-    res.send('Index page');
-
+ctrl.remove = (req, res) => {
+  res.send("Index page");
 };
 
 module.exports = ctrl;
